third_party/WebKit/LayoutTests/http/tests/media/video-load-suspend.html

<!DOCTYPE html>
<title>Test that the load eventually suspends and returns to NETWORK_IDLE.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "http://127.0.0.1:8000/resources/test.ogv";
    var watcher = new EventWatcher(t, video, ["loadstart", "suspend"]);
    watcher.wait_for(["loadstart", "suspend"]).then(t.step_func_done(function() {
        assert_equals(video.networkState, HTMLMediaElement.NETWORK_IDLE);
    }));
});
</script>
